pyproject.toml
[project]
name = "image-search"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "huggingface-hub>=0.34.4",
    "pillow>=11.3.0",
    "psycopg2>=2.9.10",
    "torch==2.8.0",
    "tqdm>=4.67.1",
    "transformers>=4.55.4",
    "numpy>=1.24.0",
    "einops>=0.8.0",
    "timm>=1.0.0",
    "torchvision>=0.20.0",
]
